Added proper attributes

+++ src/main/java/de/streamfusion/controllers/APIController.java
@@ -1,5 +1,8 @@
 package de.streamfusion.controllers;
 
+import de.streamfusion.exceptions.EmailAlreadyExistsException;
+import de.streamfusion.exceptions.NoValidEmailException;
+import de.streamfusion.exceptions.UsernameTakenException;
 import de.streamfusion.models.User;
 import de.streamfusion.models.Video;
 import de.streamfusion.services.UserService;
@@ -7,16 +10,11 @@ import de.streamfusion.services.VideoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.server.ResponseStatusException;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.Objects;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/api/v1")
@@ -30,7 +28,7 @@ public class APIController {
         this.videoService = videoService;
     }
 
-    @PostMapping(value = "/upload")
+    @PostMapping(value = "/video/upload")
     public ResponseEntity<String> uploadVideo(
         @RequestParam MultipartFile file,
         @RequestParam String title,
@@ -38,44 +36,61 @@ public class APIController {
         @RequestParam long userID
     ) {
         if (file.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "File is empty.");
+            return new ResponseEntity<>("File is empty.", HttpStatus.BAD_REQUEST);
         }
         if (title.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Title is empty.");
+            return new ResponseEntity<>("Title is empty.", HttpStatus.BAD_REQUEST);
         }
         if (description.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Description is empty.");
+            return new ResponseEntity<>("Description is empty.", HttpStatus.BAD_REQUEST);
         }
 
-        final User user = this.userService.getUserByID(userID).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "User does not exist.")
-        );
-        final String fileExtension = Objects.requireNonNull(file.getContentType()).split("/")[1];
-        if (!fileExtension.equals("mp4")) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "File is not a video.");
-        }
-        Video video = new Video();
-        video.setTitle(title);
-        video.setDescription(description);
-        video.setUser(user);
-        video.setFiletype(fileExtension);
-        this.videoService.saveVideo(video);
-
-        final File videoFile = new File("%s/data/%d/video.%s".formatted(
-                System.getProperty("user.dir"),
-                video.getID(),
-                fileExtension
-        ));
-
+        final User user;
         try {
-            if (!videoFile.getParentFile().mkdirs()) {
-                throw new IOException("Could not create parent directories.");
-            }
-            file.transferTo(videoFile);
+            user = this.userService.getUserByID(userID).orElseThrow();
+        } catch (NoSuchElementException e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            this.videoService.saveVideo(file, title, description, user);
         } catch (IOException e) {
-            this.videoService.deleteVideo(video);
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (IllegalArgumentException e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
         return new ResponseEntity<>("Video successfully uploaded.", HttpStatus.OK);
+    }
+
+    @DeleteMapping(value = "/video/delete")
+    public ResponseEntity<String> deleteVideo(
+        @RequestParam long id,
+        @RequestParam long userID
+    ) {
+        final Video video;
+        try {
+            video = this.videoService.getVideoByID(id).orElseThrow();
+        } catch (NoSuchElementException e) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        if (video.getUser().getID() != userID) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        try {
+            this.videoService.deleteVideo(video);
+        } catch (IOException e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>("Video successfully deleted.", HttpStatus.OK);
+    }
+
+    @PostMapping(value = "/register")
+    public ResponseEntity<?> registerUser(@RequestBody User user) {
+        try {
+            this.userService.addUser(user);
+        } catch (EmailAlreadyExistsException | NoValidEmailException | UsernameTakenException e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
